finish 1st kadai

+++ app/Http/Controllers/TasksController.php
@@ -15,13 +15,22 @@ class TasksController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+   public function index()
     {
-           $tasks = Task::all();
 
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            $data += $this->counts($user);
+            return view('tasks.index', $data);
+        }else {
+            return view('welcome');
+        }
     }
 
     /**
@@ -47,16 +56,15 @@ class TasksController extends Controller
      public function store(Request $request)
     {
         $this->validate($request, [
-            'status' => 'required|max:10',   // add
-            'content' => 'required|max:191',
+        'status' => 'required|max:10',   // add
+        'content' => 'required|max:191',
         ]);
-
-
+        $user = \Auth::user();
         $task = new Task;
         $task->status = $request->status;    // add
         $task->content = $request->content;
+        $task->user_id = $user->id;
         $task->save();
-
 
         return redirect('/');
     }
@@ -69,10 +77,12 @@ class TasksController extends Controller
     public function show($id)
     {
          $task = Task::find($id);
-
-        return view('tasks.show', [
+         
+         if (\Auth::user()->id === $task->user_id) {
+             return view('tasks.show', [
             'task' => $task,
         ]);
+         }
     }
 
     /**
@@ -85,9 +95,11 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
 
+        if (\Auth::user()->id === $task->user_id) {
         return view('tasks.edit', [
             'task' => $task,
         ]);
+        }
     }
 
     /**
@@ -97,7 +109,7 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-      public function update(Request $request, $id)
+      public function update(Request $request, $id) 
     {
         $this->validate($request, [
             'status' => 'required|max:10',   // add
@@ -120,11 +132,15 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+       public function destroy($id)
     {
-        $task = Task::find($id);
-        $task->delete();
+        $task = \App\Task::find($id);
 
-        return redirect('/');
+        if (\Auth::user()->id === $task->user_id) {
+            $task->delete();
+        
+            return redirect('/');
+        }
+
     }
 }
